pyproject.toml
[tool.poetry]
name = "itis_table_parse"
version = "0.1.0"
description = ""
authors = ["eale <RUSLIKPVP@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.48.0"
oauth2client = "^4.1.3"
gspread = "^5.3.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
